Revert "new protocol (chrome 62)"

+++ source/ChromeDevTools/Protocol/Chrome/Input/SynthesizePinchGestureCommand.cs
@@ -1,6 +1,5 @@
 using MasterDevs.ChromeDevTools;
 using Newtonsoft.Json;
-using System;
 using System.Collections.Generic;
 
 namespace MasterDevs.ChromeDevTools.Protocol.Chrome.Input
@@ -15,11 +14,11 @@ namespace MasterDevs.ChromeDevTools.Protocol.Chrome.Input
 		/// <summary>
 		/// Gets or sets X coordinate of the start of the gesture in CSS pixels.
 		/// </summary>
-		public double X { get; set; }
+		public long X { get; set; }
 		/// <summary>
 		/// Gets or sets Y coordinate of the start of the gesture in CSS pixels.
 		/// </summary>
-		public double Y { get; set; }
+		public long Y { get; set; }
 		/// <summary>
 		/// Gets or sets Relative scale factor after zooming (>1.0 zooms in, <1.0 zooms out).
 		/// </summary>
